requirements.txt
aioredis==2.0.1
arxiv==1.4.2
async-timeout==4.0.2
attrs==21.4.0
autosemver==0.5.5
Babel==2.9.1
beautifulsoup4==4.10.0
black==22.1.0
blis==0.7.5
boto3==1.21.18
botocore==1.24.18
bs4==0.0.1
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
cleo==1.0.0a4
click==8.0.3
crashtest==0.3.1
cryptography==36.0.1
cymem==2.0.6
decorator==5.1.1
Deprecated==1.2.13
docopt==0.6.2
dulwich==0.19.16
feedparser==6.0.8
flasgger==0.9.5
Flask==2.0.3
greenlet==1.1.2
gssapi==1.7.3
hdfs==2.6.0
hiredis==2.0.0
idna==3.3
importlib-resources==5.4.0
inspire-utils==3.0.24
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.4.0
kerberos==1.3.1
krb5==0.3.0
langcodes==3.3.0
libmagic==1.0
lxml==4.7.1
MarkupSafe==2.0.1
mistune==2.0.2
murmurhash==1.0.6
mypy-extensions==0.4.3
nameparser==0.5.8
nltk==3.6.7
numpy==1.22.1
Owlready2==0.36
packaging==21.3
pathspec==0.9.0
pathy==0.6.1
pdfminer.six==20201018
platformdirs==2.5.1
pptree==3.1
preshed==3.0.6
pycparser==2.21
pydantic==1.8.2
pygrobid==0.1.6
pylev==1.4.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyrsistent==0.18.1
pyspnego==0.5.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-magic==0.4.25
python-ulid==1.0.3
pytz==2021.3
PyYAML==6.0
redis==4.1.2
redis-om==0.0.17
regex==2022.1.18
requests==2.27.1
requests-kerberos==0.14.0
s3transfer==0.5.2
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
sortedcontainers==2.4.0
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.31
srsly==2.4.2
thinc==8.0.13
thread6==0.2.0
tomli==2.0.1
tqdm==4.62.3
typer==0.4.0
types-redis==4.1.11
types-six==1.16.10
typing-extensions==4.0.1
Unidecode==1.3.2
urllib3==1.26.8
waitress==2.1.1
wasabi==0.9.0
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
